fix(template): remove scene namespace
