--- Templates/BEScene2.xctemplate/___VARIABLE_scene___.ViewController.swift
@@ -1,47 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//
-
-import BEPureLayout
-
-extension ___VARIABLE_scene___ {
-    final class ViewController: BEScene {
-        // MARK: - Dependencies
-
-        private let viewModel: ViewModel
-
-        // MARK: - Initializer
-
-        init(viewModel: ViewModel) {
-            self.viewModel = viewModel
-            super.init()
-        }
-
-        // MARK: - Methods
-
-        override func build() -> UIView {
-            RootView(viewModel: ViewModel)
-        }
-
-        override func bind() {
-            super.bind()
-            viewModel.output.navigatableScene
-                .emit(onNext: { [weak self] in self?.navigate(to: $0) })
-                .disposed(by: disposeBag)
-        }
-
-        // MARK: - Navigation
-
-        private func navigate(to scene: NavigatableScene?) {
-            switch scene {
-            case .detail:
-                <#code#>
-            case .none:
-                break
-            }
-        }
-    }
-}
