+++ Templates/BEScene2.xctemplate/___VARIABLE_scene___NavigatableScene.swift
@@ -7,8 +7,6 @@
 
 import Foundation
 
-enum ___VARIABLE_scene___ {
-    enum NavigatableScene {
-        case detail
-    }
+enum ___VARIABLE_scene___NavigatableScene {
+    case detail
 }
